Enable absolute snapshot path
I have a @snapshot subvolume at mountpoint /snapshots that isn't mounted per default for two reasons:

* obscure the existence of snapshots, just in case an attacker has root access and could delete subvolumes
* don't pollute the dir to be snapshotted with unnecessary folders like .snapshot

I want to store the snapshots there. The previous version would only support storing the snapshots in a child path of the subvolume to be snapshotted.

The destination path isn't required to be mounted. But it must exist as a mountpoint in /etc/fstab.